refactor: use useProduct hook on Products page
